windows-optimize-smb-dir-search-if-no-wildcard-20061217
When performing a SMB FindFirst/FindNext/FindClose operation if there
are no wildcards involved, we can optimize the case and turn it from
O(n) to O(1) where 'n' is the number of entries in the directory.
This can be done by performing a cm_Lookup() and if it succeeds,
constructing the appropriate response instead of parsing the contents
of each buffer associated with the directory looking for matches.

Without this optimization, FindFirst operations on directories containing
thousands of entries can take a large number of seconds to complete.

+++ src/WINNT/afsd/cm_conn.c
@@ -22,7 +22,7 @@
 
 osi_rwlock_t cm_connLock;
 
-long RDRtimeout = CM_CONN_DEFAULTRDRTIMEOUT;
+DWORD RDRtimeout = CM_CONN_DEFAULTRDRTIMEOUT;
 unsigned short ConnDeadtimeout = CM_CONN_CONNDEADTIME;
 unsigned short HardDeadtimeout = CM_CONN_HARDDEADTIME;
 
@@ -86,7 +86,7 @@ void cm_InitConn(void)
 	    RegCloseKey(parmKey);
 	}
 
-	afsi_log("lanmanworkstation : SessTimeout %d", RDRtimeout);
+	afsi_log("lanmanworkstation : SessTimeout %u", RDRtimeout);
 	if (ConnDeadtimeout == 0)
 	    ConnDeadtimeout = (unsigned short) (RDRtimeout / 2);
 	afsi_log("ConnDeadTimeout is %d", ConnDeadtimeout);
